Spark: Refactor action for rewriting manifests
