+++ spark/src/main/java/org/apache/iceberg/actions/Actions.java
@@ -24,6 +24,7 @@ import org.apache.iceberg.common.DynConstructors;
 import org.apache.iceberg.common.DynMethods;
 import org.apache.iceberg.spark.actions.BaseExpireSnapshotsSparkAction;
 import org.apache.iceberg.spark.actions.BaseRemoveOrphanFilesSparkAction;
+import org.apache.iceberg.spark.actions.BaseRewriteManifestsSparkAction;
 import org.apache.spark.sql.SparkSession;
 
 public class Actions {
@@ -74,7 +75,8 @@ public class Actions {
   }
 
   public RewriteManifestsAction rewriteManifests() {
-    return new RewriteManifestsAction(spark, table);
+    BaseRewriteManifestsSparkAction delegate = new BaseRewriteManifestsSparkAction(spark, table);
+    return new RewriteManifestsAction(delegate);
   }
 
   public RewriteDataFilesAction rewriteDataFiles() {
